+++ src/lib/unlang/xlat_priv.h
@@ -162,7 +162,7 @@ struct xlat_exp_s {
 	union {
 		struct {
 			xlat_exp_head_t	*group;		//!< children of a group
-			uint8_t		hoist : 1;	//!< it's a group, but we need to hoist the results
+			unsigned int   	hoist : 1;	//!< it's a group, but we need to hoist the results
 		};
 
 		/** An tmpl_t reference
@@ -188,10 +188,10 @@ struct xlat_exp_s {
 struct xlat_exp_head_s {
 	fr_dlist_head_t		dlist;
 	xlat_flags_t		flags;		//!< Flags that control resolution and evaluation.
-	uint8_t			instantiated : 1;  //!< temporary flag until we fix more things
-	uint8_t			is_argv : 1;	//!< this thing holds function arguments
-	uint8_t			cursor : 1;	//!< otherwise it's too hard to pass xlat_arg_parser_t to the evaluation function.
-	uint8_t			is_attr : 1;	//!< the argument is an attribute reference
+	unsigned int		instantiated : 1;  //!< temporary flag until we fix more things
+	unsigned int		is_argv : 1;	//!< this thing holds function arguments
+	unsigned int		cursor : 1;	//!< otherwise it's too hard to pass xlat_arg_parser_t to the evaluation function.
+	unsigned int		is_attr : 1;	//!< the argument is an attribute reference
 
 #ifndef NDEBUG
 	char const * _CONST	file;		//!< File where the xlat was allocated.
